Delete CODEOWNERS
Apparently no one would like to receive automatic assignment of all PRs.
Especially not Zach and Liv.

For reference, multiple lines are undesired syntax:
Multiple people must be on the same line of CODEOWNERS, otherwise only
the last line of a given pattern (*) is used. We don't want to force
*everything* on Zach.