StateEligiblity.js: Eligibility string enhancements (#119)
* StateEligibility.js: Refactor to editable strings

Use an Array of Strings (actually an Array of Objects of Strings) to
generate the <ListItem/>s, instead of manually entering all of them.
Now editing the criteria is much simpler and more centralized.

Remove the EligibilityGroupItem abstraction for ListItem, it was
unnecessary and seperates the icon from the text, making it harder to
parametrize the icon in the future.

Handle the goofiness that sometimes we have a title <div> and
sometimes we do not.

Unfortunately we have to add keys to dynamically generated react
fragments to quell warnings about duplicate keys, which means we have to use the
<React.Fragment key="xyz"> ... </React.Fragment>
syntax rather than the more common abbreviated
<> ... </>
syntax. Hopefully this isn't too confusing.

* StateEligibility.js: Use PeopleIcon instead of CheckIcon

As noted by me and confirmed by Hal, the CheckIcon had an affordance
for completion, the implication that you're supposed to do all the
checkable items. Although it's not seriously contemplated that anyone
would actually *think* that, it's the wrong application for a checkbox
icon.

Although we originally discussed using a bullet icon, material UI
doesn't seem to have a good choice -- there is
RadioButtonUncheckedIcon but it's a fairly large circle.

Instead, let's use icons that actually represent groups of people,
which is what we are listing anyhow. That means...PeopleIcon!

* StateEligibility.js: Update teachers' URL

Not final. This is a better URL than the generic one that does not
explain the policy, but it's still a press release and not a really
good simple explanation. But that's lacking, so whacha-gonna-do?

* StateEligibility.js: Use MassGov's wording for teacher category

Specifically, remove "Pre-K" as that language does not appear in
the mass.gov announcement we link to, even though such people probably
fall under "child care workers," which does.

This wording originally came in via
cf528d1c0d95db7d5b54e9bb3b224b09bc026f07.

Co-authored-by: Rob Harris <rob@harcod.com>

Co-authored-by: Rob Harris <rob@harcod.com>