No longer skips test (#144)
* No longer skips test

Test: "when the api endpoint can not be reached"

Before: Was failing because there are multiple "role=alert" on the screen when the site's AlertBanner is displayed.

Now: This test checks for the existence of "Unexpected Internal Error" which is the title of the ErrorMessageAlert

* remove useless comment

* filters were failing on undefined coordinates

* No longer skips test associated with filtering.

The test was update to click on the "Only Show locations with available appointments" checkbox, and the "Update List" button.

* find filter buttons by data-testid instead of text for i18n

* rollback to prior version

* data-testid unique in both FilterPanel (mobile & desktop)