env test changed
The env test was not working as expected because "true" is a string (not a boolean) in env files.

Changed the test from the negative of "IGNORE_DEVTEST_JSON" to the affirmative "USE_DEVTEST_JSON".  The production server will not have this set.  Also, there is no longer a need for ".env.test" to set ignore.