Testing front-end with local data (#145)
* No longer skips test

Test: "when the api endpoint can not be reached"

Before: Was failing because there are multiple "role=alert" on the screen when the site's AlertBanner is displayed.

Now: This test checks for the existence of "Unexpected Internal Error" which is the title of the ErrorMessageAlert

* remove useless comment

* No longer skips test associated with filtering.

The test was update to click on the "Only Show locations with available appointments" checkbox, and the "Update List" button.

* Handles case where latitude and longitude coordinates for a location are not in the data file.

Some of our archived data doesn't contain coordinates at all.

* Removed unused code that was refactored to "appointmentData.service.js"

* Test with local data

Development only change

Allows the front-end to specify test data from a file instead of fetching live data from production.

The test file is in test/devtest.json.

If the file doesn't exist, then it fetches the production data.

* Timestamp from test data used as "now" reference

* added .env.test file

Sets an environment variable while running 'npm run test' that keeps the devtest.json file from loading and messing up the mocked test data.

* replaced skipped test caused by bad merge

* prefer 'import' over 'require'

* removed testing code

* removed unused test file

* gitignore devtest files

* env test changed

The env test was not working as expected because "true" is a string (not a boolean) in env files.

Changed the test from the negative of "IGNORE_DEVTEST_JSON" to the affirmative "USE_DEVTEST_JSON".  The production server will not have this set.  Also, there is no longer a need for ".env.test" to set ignore.

* README.md instructions for using test feature

* merged changes and removed console

Co-authored-by: Olivia Adams <liv.gust@gmail.com>